backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
websockets>=12.0
pydantic>=2.6.0
demucs>=4.0.1
torch>=2.1.0
torchaudio>=2.1.0
yt-dlp>=2024.1.0
soundfile>=0.12.1
numpy>=1.24.0
librosa>=0.10.2
aiofiles>=23.2.1
requests
pytest
httpx
PyJWT>=2.8.0
